Fix Assert.Equals errors by casting to (ulong)

+++ Source/EasyNetQ.Tests/TimeoutStrategyTest.cs
@@ -18,14 +18,14 @@ namespace EasyNetQ.Tests
         public void TestWhenMessagetWithAttribute()
         {
             var timeoutStrategy = new TimeoutStrategy(new ConnectionConfiguration {Timeout = 10});
-            Assert.Equal(90, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithTimeoutAttribute)));
+            Assert.Equal((ulong)90, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithTimeoutAttribute)));
         }
 
         [Fact]
         public void TestWhenPersistentMessagesIsFalse()
         {
             var timeoutStrategy = new TimeoutStrategy(new ConnectionConfiguration { Timeout = 10 });
-            Assert.Equal(10, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithoutTimeoutAttribute)));
+            Assert.Equal((ulong)10, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithoutTimeoutAttribute)));
         }
     }
 }
